Improved documentation

+++ range_queries/segtree.cpp
@@ -1,7 +1,25 @@
+/**
+ * @file    segtree.cpp
+ * @brief   Implementation of [Segment Tree]
+ *          (https://en.wikipedia.org/wiki/Segment_tree) data structure
+ *
+ * @details
+ * A segment tree, also known as a statistic tree, is a tree data structure used
+ * for storing information about intervals, or segments. Its classical version
+ * allows querying which of the stored segments contain a given point, but our
+ * modification allows us to perform (query) any binary operation on any range
+ * in the array in O(logN) time. Here, we have used addition (+).
+ * For range updates, we have used lazy propagation.
+ *
+ * * Space Complexity : O(NlogN) \n
+ * * Build Time Complexity : O(NlogN) \n
+ * * Query Time Complexity : O(logN) \n
+ */
+
 #include <cassert>   /// for assert
 #include <cmath>     /// for log2
-#include <iostream>  /// for io
-#include <vector>    /// for vector
+#include <iostream>  /// for IO operations
+#include <vector>    /// for std::vector
 
 /**
  * @brief   Constructs the initial segment tree
@@ -13,14 +31,14 @@
  * @param   pos index of segtree to fill (eg. root node)
  * @returns void
  */
-void ConsTree(const std::vector<int> &arr, std::vector<int> *segtree, int low,
-              int high, int pos) {
+void ConsTree(const std::vector<int64_t> &arr, std::vector<int64_t> *segtree,
+              uint64_t low, uint64_t high, uint64_t pos) {
     if (low == high) {
         (*segtree)[pos] = arr[low];
         return;
     }
 
-    int mid = (low + high) / 2;
+    uint64_t mid = (low + high) / 2;
     ConsTree(arr, segtree, low, mid, 2 * pos + 1);
     ConsTree(arr, segtree, mid + 1, high, 2 * pos + 2);
     (*segtree)[pos] = (*segtree)[2 * pos + 1] + (*segtree)[2 * pos + 2];
@@ -36,10 +54,11 @@ void ConsTree(const std::vector<int> &arr, std::vector<int> *segtree, int low,
  * @param   low lower index of query for this function call
  * @param   high higher index of query for this function call
  * @param   pos index of segtree to consider (eg. root node)
- * @return  int result of the range query for this function call
+ * @return  result of the range query for this function call
  */
-int query(std::vector<int> *segtree, std::vector<int> *lazy, int qlow,
-          int qhigh, int low, int high, int pos) {
+int64_t query(std::vector<int64_t> *segtree, std::vector<int64_t> *lazy,
+              uint64_t qlow, uint64_t qhigh, uint64_t low, uint64_t high,
+              uint64_t pos) {
     if (low > high || qlow > high || low > qhigh) {
         return 0;
     }
@@ -58,7 +77,7 @@ int query(std::vector<int> *segtree, std::vector<int> *lazy, int qlow,
         return (*segtree)[pos];
     }
 
-    int mid = (low + high) / 2;
+    uint64_t mid = (low + high) / 2;
 
     return query(segtree, lazy, qlow, qhigh, low, mid, 2 * pos + 1) +
            query(segtree, lazy, qlow, qhigh, mid + 1, high, 2 * pos + 2);
@@ -77,8 +96,9 @@ int query(std::vector<int> *segtree, std::vector<int> *lazy, int qlow,
  * @param   pos index of segtree to consider (eg. root node)
  * @returns void
  */
-void update(std::vector<int> *segtree, std::vector<int> *lazy, int start,
-            int end, int delta, int low, int high, int pos) {
+void update(std::vector<int64_t> *segtree, std::vector<int64_t> *lazy,
+            int64_t start, int64_t end, int64_t delta, uint64_t low,
+            uint64_t high, uint64_t pos) {
     if (low > high) {
         return;
     }
@@ -108,7 +128,7 @@ void update(std::vector<int> *segtree, std::vector<int> *lazy, int start,
         return;
     }
 
-    int mid = (low + high) / 2;
+    uint64_t mid = (low + high) / 2;
 
     update(segtree, lazy, start, end, delta, low, mid, 2 * pos + 1);
     update(segtree, lazy, start, end, delta, mid + 1, high, 2 * pos + 2);
@@ -121,10 +141,10 @@ void update(std::vector<int> *segtree, std::vector<int> *lazy, int start,
  * @returns void
  */
 static void test() {
-    int max = static_cast<int>(2 * pow(2, ceil(log2(7))) - 1);
+    int64_t max = static_cast<int64_t>(2 * pow(2, ceil(log2(7))) - 1);
     assert(max == 15);
 
-    std::vector<int> arr{1, 2, 3, 4, 5, 6, 7}, lazy(max), segtree(max);
+    std::vector<int64_t> arr{1, 2, 3, 4, 5, 6, 7}, lazy(max), segtree(max);
     ConsTree(arr, &segtree, 0, 7 - 1, 0);
 
     assert(query(&segtree, &lazy, 1, 5, 0, 7 - 1, 0) == 2 + 3 + 4 + 5 + 6);
@@ -142,15 +162,15 @@ static void test() {
  * @return  0 on exit
  */
 int main() {
-    test();
+    test();  // run self-test implementations
 
     std::cout << "Enter number of elements: ";
 
-    int n = 0;
+    uint64_t n = 0;
     std::cin >> n;
 
-    int max = static_cast<int>(2 * pow(2, ceil(log2(n))) - 1);
-    std::vector<int> arr(n), lazy(max), segtree(max);
+    uint64_t max = static_cast<uint64_t>(2 * pow(2, ceil(log2(n))) - 1);
+    std::vector<int64_t> arr(n), lazy(max), segtree(max);
 
     int choice = 0;
     std::cout << "\nDo you wish to enter each number?:\n"
@@ -178,13 +198,13 @@ int main() {
         if (choice == 1) {
             std::cout << "Enter 1-indexed lower bound, upper bound & value:\n";
 
-            int p = 1, q = 1, v = 0;
+            uint64_t p = 1, q = 1, v = 0;
             std::cin >> p >> q >> v;
             update(&segtree, &lazy, p - 1, q - 1, v, 0, n - 1, 0);
         } else if (choice == 2) {
             std::cout << "Enter 1-indexed lower bound & upper bound:\n";
 
-            int p = 1, q = 1;
+            uint64_t p = 1, q = 1;
             std::cin >> p >> q;
             std::cout << query(&segtree, &lazy, p - 1, q - 1, 0, n - 1, 0);
             std::cout << "\n";
